Non adaptative gaussian model working

+++ src/operations/morphological_operations.py
@@ -3,8 +3,8 @@ import cv2
 
 
 def opening(im: np.ndarray, kernel_side: int) -> np.ndarray:
-    return cv2.morphologyEx(im, cv2.MORPH_OPEN, np.ones((kernel_side, kernel_side)))
+    return cv2.morphologyEx(im, cv2.MORPH_OPEN, np.ones((kernel_side, kernel_side), dtype=np.uint8))
 
 
 def closing(im: np.ndarray, kernel_side: int) -> np.ndarray:
-    return cv2.morphologyEx(im, cv2.MORPH_CLOSE, np.ones((kernel_side, kernel_side)))
+    return cv2.morphologyEx(im, cv2.MORPH_CLOSE, np.ones((kernel_side, kernel_side), dtype=np.uint8))
